Adds namespace for helper functions
