+++ composer.json
@@ -22,7 +22,7 @@
             "FriendsOfHyperf\\Jet\\": "src/"
         },
         "files": [
-            "src/helpers.php"
+            "src/Functions.php"
         ]
     },
     "autoload-dev": {
