+++ src/PathGenerator/PathGenerator.php
@@ -14,6 +14,8 @@ namespace FriendsOfHyperf\Jet\PathGenerator;
 
 use FriendsOfHyperf\Jet\Contract\PathGeneratorInterface;
 
+use function FriendsOfHyperf\Jet\str_snake;
+
 class PathGenerator implements PathGeneratorInterface
 {
     public function generate(string $service, string $method): string
